+++ packages/omg-js-util/src/hexPrefix.js
